Fix npm run test:unit
