+++ src/apps/companies/apps/activity-feed/__test__/controllers.test.js
@@ -12,10 +12,9 @@ const essDetails = require('../../../../../../test/sandbox/fixtures/v4/activity-
 const contactMock = require('../../../../../../test/sandbox/fixtures/v3/contact/contact-by-id-uk.json')
 
 const {
-  // DATA_HUB_ACTIVITY,
-  // EXTERNAL_ACTIVITY,
-  // DATA_HUB_AND_EXTERNAL_ACTIVITY,
-  // DATA_HUB_AND_AVENTRI_ACTIVITY,
+  EXTERNAL_ACTIVITY,
+  DATA_HUB_AND_EXTERNAL_ACTIVITY,
+  DATA_HUB_AND_AVENTRI_ACTIVITY,
   EVENT_AVENTRI_ATTENDEES_STATUS,
   EVENT_ATTENDEES_SORT_OPTIONS,
   EVENT_ATTENDEES_STATUS,
@@ -28,7 +27,7 @@ const {
   filterContactListOnEmail,
 } = require('../controllers')
 const { has, get } = require('lodash')
-// const { sortCriteria } = require('../es-queries/sortCriteria')
+const { sortCriteria } = require('../es-queries/sortCriteria')
 
 describe('Activity feed controllers', () => {
   let fetchActivityFeedStub,
@@ -77,7 +76,6 @@ describe('Activity feed controllers', () => {
             middlewareParameters = buildMiddlewareParameters({
               company: companyMock,
               requestQuery: {
-                activityTypeFilter: 'dataHubAndExternalActivity',
                 showDnbHierarchy: false,
               },
               user: {
@@ -94,31 +92,106 @@ describe('Activity feed controllers', () => {
 
           it('should call fetchActivityFeed with correct params when retrieving aventri attendees', async () => {
             const expectedEsQuery = {
+              from: 0,
+              size: 20,
+              sort: sortCriteria('desc'),
               query: {
                 bool: {
-                  must: [
-                    {
-                      term: {
-                        'object.type': 'dit:aventri:Attendee',
-                      },
+                  filter: {
+                    bool: {
+                      should: [
+                        {
+                          bool: {
+                            must: [
+                              {
+                                terms: {
+                                  'object.type': DATA_HUB_AND_EXTERNAL_ACTIVITY,
+                                },
+                              },
+                              {
+                                terms: {
+                                  'object.attributedTo.id': [
+                                    'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
+                                  ],
+                                },
+                              },
+                            ],
+                          },
+                        },
+                        {
+                          bool: {
+                            must: [
+                              {
+                                term: {
+                                  'object.type':
+                                    'dit:directoryFormsApi:Submission',
+                                },
+                              },
+                              {
+                                term: {
+                                  'object.attributedTo.type':
+                                    'dit:directoryFormsApi:SubmissionAction:gov-notify-email',
+                                },
+                              },
+                              {
+                                term: {
+                                  'object.url':
+                                    '/contact/export-advice/comment/',
+                                },
+                              },
+                              {
+                                terms: {
+                                  'actor.dit:emailAddress': [
+                                    'fred@acme.org',
+                                    'fred@acme.org',
+                                    'fred@acme.org',
+                                    'byvanuwenu@yahoo.com',
+                                  ],
+                                },
+                              },
+                            ],
+                          },
+                        },
+                        {
+                          bool: {
+                            must: [
+                              { term: { 'object.type': 'dit:aventri:Event' } },
+                              {
+                                terms: {
+                                  id: [
+                                    'dit:aventri:Event:1:Create',
+                                    'dit:aventri:Event:2:Create',
+                                  ],
+                                },
+                              },
+                            ],
+                          },
+                        },
+                        {
+                          bool: {
+                            must: [
+                              {
+                                term: {
+                                  'object.attributedTo.id':
+                                    'dit:directoryFormsApi:SubmissionType:export-support-service',
+                                },
+                              },
+                              {
+                                terms: {
+                                  'actor.dit:emailAddress': [
+                                    'fred@acme.org',
+                                    'fred@acme.org',
+                                    'fred@acme.org',
+                                    'byvanuwenu@yahoo.com',
+                                  ],
+                                },
+                              },
+                            ],
+                          },
+                        },
+                      ],
                     },
-                    {
-                      terms: {
-                        'object.dit:emailAddress': [
-                          'fred@acme.org',
-                          'fred@acme.org',
-                          'fred@acme.org',
-                          'byvanuwenu@yahoo.com',
-                        ],
-                      },
-                    },
-                  ],
-                },
-              },
-              sort: {
-                'object.updated': {
-                  order: 'desc',
-                  unmapped_type: 'date',
+                  },
                 },
               },
             }
@@ -128,122 +201,6 @@ describe('Activity feed controllers', () => {
               expectedEsQuery
             )
           })
-
-          // it('should call fetchActivityFeed with the right params', async () => {
-          //   const expectedEsQuery = {
-          //     from: 0,
-          //     size: 20,
-          //     sort: sortCriteria('desc'),
-          //     query: {
-          //       bool: {
-          //         filter: {
-          //           bool: {
-          //             should: [
-          //               {
-          //                 bool: {
-          //                   must: [
-          //                     {
-          //                       terms: {
-          //                         'object.type': DATA_HUB_AND_EXTERNAL_ACTIVITY,
-          //                       },
-          //                     },
-          //                     {
-          //                       terms: {
-          //                         'object.attributedTo.id': [
-          //                           'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
-          //                         ],
-          //                       },
-          //                     },
-          //                   ],
-          //                 },
-          //               },
-          //               {
-          //                 bool: {
-          //                   must: [
-          //                     {
-          //                       term: {
-          //                         'object.type':
-          //                           'dit:directoryFormsApi:Submission',
-          //                       },
-          //                     },
-          //                     {
-          //                       term: {
-          //                         'object.attributedTo.type':
-          //                           'dit:directoryFormsApi:SubmissionAction:gov-notify-email',
-          //                       },
-          //                     },
-          //                     {
-          //                       term: {
-          //                         'object.url':
-          //                           '/contact/export-advice/comment/',
-          //                       },
-          //                     },
-          //                     {
-          //                       terms: {
-          //                         'actor.dit:emailAddress': [
-          //                           'fred@acme.org',
-          //                           'fred@acme.org',
-          //                           'fred@acme.org',
-          //                           'byvanuwenu@yahoo.com',
-          //                         ],
-          //                       },
-          //                     },
-          //                   ],
-          //                 },
-          //               },
-          //               {
-          //                 bool: {
-          //                   must: [
-          //                     {
-          //                       term: {
-          //                         'object.type': 'dit:aventri:Event',
-          //                       },
-          //                     },
-          //                     {
-          //                       terms: {
-          //                         id: [
-          //                           'dit:aventri:Event:1:Create',
-          //                           'dit:aventri:Event:2:Create',
-          //                         ],
-          //                       },
-          //                     },
-          //                   ],
-          //                 },
-          //               },
-          //               {
-          //                 bool: {
-          //                   must: [
-          //                     {
-          //                       term: {
-          //                         'object.attributedTo.id':
-          //                           'dit:directoryFormsApi:SubmissionType:export-support-service',
-          //                       },
-          //                     },
-          //                     {
-          //                       terms: {
-          //                         'actor.dit:emailAddress': [
-          //                           'fred@acme.org',
-          //                           'fred@acme.org',
-          //                           'fred@acme.org',
-          //                           'byvanuwenu@yahoo.com',
-          //                         ],
-          //                       },
-          //                     },
-          //                   ],
-          //                 },
-          //               },
-          //             ],
-          //           },
-          //         },
-          //       },
-          //     },
-          //   }
-
-          //   expect(fetchActivityFeedStub).to.be.calledWith(
-          //     middlewareParameters.reqMock,
-          //     expectedEsQuery
-          //   )
-          // })
         })
       }
     )
@@ -253,8 +210,8 @@ describe('Activity feed controllers', () => {
         middlewareParameters = buildMiddlewareParameters({
           company: companyMock,
           requestQuery: {
-            activityTypeFilter: 'myActivity',
             showDnbHierarchy: false,
+            dit_participants__adviser: [123],
           },
           user: {
             id: 123,
@@ -268,146 +225,140 @@ describe('Activity feed controllers', () => {
         )
       })
 
-      // it('should call fetchActivityFeed with a user id', async () => {
-      //   const expectedEsQuery = {
-      //     from: 0,
-      //     size: 20,
-      //     sort: sortCriteria('desc'),
-      //     query: {
-      //       bool: {
-      //         must: [
-      //           {
-      //             terms: {
-      //               'object.type': DATA_HUB_ACTIVITY,
-      //             },
-      //           },
-      //           {
-      //             term: {
-      //               'object.attributedTo.id': 'dit:DataHubAdviser:123',
-      //             },
-      //           },
-      //           {
-      //             terms: {
-      //               'object.attributedTo.id': [
-      //                 'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
-      //               ],
-      //             },
-      //           },
-      //         ],
-      //       },
-      //     },
-      //   }
-
-      //   expect(fetchActivityFeedStub).to.be.calledWith(
-      //     middlewareParameters.reqMock,
-      //     expectedEsQuery
-      //   )
-      // })
-    })
-
-    context('when filtering on "Data Hub activity" for a company', () => {
-      before(async () => {
-        middlewareParameters = buildMiddlewareParameters({
-          company: companyMock,
-          requestQuery: {
-            activityTypeFilter: 'dataHubActivity',
-            showDnbHierarchy: false,
+      it('should call fetchActivityFeed with a user id', async () => {
+        const expectedEsQuery = {
+          from: 0,
+          size: 20,
+          sort: sortCriteria('desc'),
+          query: {
+            bool: {
+              filter: {
+                bool: {
+                  should: [
+                    {
+                      bool: {
+                        must: [
+                          {
+                            terms: {
+                              'object.type': DATA_HUB_AND_EXTERNAL_ACTIVITY,
+                            },
+                          },
+                          {
+                            terms: {
+                              'object.attributedTo.id': [
+                                'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
+                              ],
+                            },
+                          },
+                          {
+                            term: {
+                              'object.attributedTo.id':
+                                'dit:DataHubAdviser:123',
+                            },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      bool: {
+                        must: [
+                          {
+                            term: {
+                              'object.type': 'dit:directoryFormsApi:Submission',
+                            },
+                          },
+                          {
+                            term: {
+                              'object.attributedTo.type':
+                                'dit:directoryFormsApi:SubmissionAction:gov-notify-email',
+                            },
+                          },
+                          {
+                            term: {
+                              'object.url': '/contact/export-advice/comment/',
+                            },
+                          },
+                          {
+                            terms: {
+                              'actor.dit:emailAddress': [
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'byvanuwenu@yahoo.com',
+                              ],
+                            },
+                          },
+                          {
+                            term: {
+                              'object.attributedTo.id':
+                                'dit:DataHubAdviser:123',
+                            },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      bool: {
+                        must: [
+                          { term: { 'object.type': 'dit:aventri:Event' } },
+                          {
+                            terms: {
+                              id: [
+                                'dit:aventri:Event:1:Create',
+                                'dit:aventri:Event:2:Create',
+                              ],
+                            },
+                          },
+                          {
+                            term: {
+                              'object.attributedTo.id':
+                                'dit:DataHubAdviser:123',
+                            },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      bool: {
+                        must: [
+                          {
+                            term: {
+                              'object.attributedTo.id':
+                                'dit:directoryFormsApi:SubmissionType:export-support-service',
+                            },
+                          },
+                          {
+                            terms: {
+                              'actor.dit:emailAddress': [
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'byvanuwenu@yahoo.com',
+                              ],
+                            },
+                          },
+                          {
+                            term: {
+                              'object.attributedTo.id':
+                                'dit:DataHubAdviser:123',
+                            },
+                          },
+                        ],
+                      },
+                    },
+                  ],
+                },
+              },
+            },
           },
-          user: {
-            id: 123,
-          },
-        })
+        }
 
-        await controllers.fetchActivityFeedHandler(
+        expect(fetchActivityFeedStub).to.be.calledWith(
           middlewareParameters.reqMock,
-          middlewareParameters.resMock,
-          middlewareParameters.nextSpy
+          expectedEsQuery
         )
       })
-
-      // it('should call fetchActivityFeed with the right params', async () => {
-      //   const expectedEsQuery = {
-      //     from: 0,
-      //     size: 20,
-      //     sort: sortCriteria('desc'),
-      //     query: {
-      //       bool: {
-      //         filter: {
-      //           bool: {
-      //             should: [
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       terms: {
-      //                         'object.type': DATA_HUB_AND_AVENTRI_ACTIVITY,
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         'object.attributedTo.id': [
-      //                           'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       term: {
-      //                         'object.type': 'dit:aventri:Event',
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         id: [
-      //                           'dit:aventri:Event:1:Create',
-      //                           'dit:aventri:Event:2:Create',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       term: {
-      //                         'object.attributedTo.id':
-      //                           'dit:directoryFormsApi:SubmissionType:export-support-service',
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         'actor.dit:emailAddress': [
-      //                           'fred@acme.org',
-      //                           'fred@acme.org',
-      //                           'fred@acme.org',
-      //                           'byvanuwenu@yahoo.com',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //             ],
-      //           },
-      //         },
-      //       },
-      //     },
-      //   }
-
-      //   expect(fetchActivityFeedStub).to.be.calledWith(
-      //     middlewareParameters.reqMock,
-      //     expectedEsQuery
-      //   )
-      // })
     })
-
     context('when filtering on "External activity" with the aventri', () => {
       before(async () => {
         middlewareParameters = buildMiddlewareParameters({
@@ -428,97 +379,100 @@ describe('Activity feed controllers', () => {
         )
       })
 
-      // it('should call fetchActivityFeed with the right params', async () => {
-      //   const expectedEsQuery = {
-      //     from: 0,
-      //     size: 20,
-      //     sort: sortCriteria('desc'),
-      //     query: {
-      //       bool: {
-      //         filter: {
-      //           bool: {
-      //             should: [
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       terms: {
-      //                         'object.type': EXTERNAL_ACTIVITY,
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         'object.attributedTo.id': [
-      //                           'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       term: {
-      //                         'object.type': 'dit:directoryFormsApi:Submission',
-      //                       },
-      //                     },
-      //                     {
-      //                       term: {
-      //                         'object.attributedTo.type':
-      //                           'dit:directoryFormsApi:SubmissionAction:gov-notify-email',
-      //                       },
-      //                     },
-      //                     {
-      //                       term: {
-      //                         'object.url': '/contact/export-advice/comment/',
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         'actor.dit:emailAddress': [
-      //                           'fred@acme.org',
-      //                           'fred@acme.org',
-      //                           'fred@acme.org',
-      //                           'byvanuwenu@yahoo.com',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       term: {
-      //                         'object.type': 'dit:aventri:Event',
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         id: [
-      //                           'dit:aventri:Event:1:Create',
-      //                           'dit:aventri:Event:2:Create',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //             ],
-      //           },
-      //         },
-      //       },
-      //     },
-      //   }
+      it('should call fetchActivityFeed with the right params', async () => {
+        const expectedEsQuery = {
+          from: 0,
+          size: 20,
+          sort: sortCriteria('desc'),
+          query: {
+            bool: {
+              filter: {
+                bool: {
+                  should: [
+                    {
+                      bool: {
+                        must: [
+                          {
+                            terms: {
+                              'object.type': [
+                                ...DATA_HUB_AND_AVENTRI_ACTIVITY,
+                                ...EXTERNAL_ACTIVITY,
+                              ],
+                            },
+                          },
+                          {
+                            terms: {
+                              'object.attributedTo.id': [
+                                'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
+                              ],
+                            },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      bool: {
+                        must: [
+                          {
+                            term: {
+                              'object.type': 'dit:directoryFormsApi:Submission',
+                            },
+                          },
+                          {
+                            term: {
+                              'object.attributedTo.type':
+                                'dit:directoryFormsApi:SubmissionAction:gov-notify-email',
+                            },
+                          },
+                          {
+                            term: {
+                              'object.url': '/contact/export-advice/comment/',
+                            },
+                          },
+                          {
+                            terms: {
+                              'actor.dit:emailAddress': [
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'byvanuwenu@yahoo.com',
+                              ],
+                            },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      bool: {
+                        must: [
+                          {
+                            term: {
+                              'object.type': 'dit:aventri:Event',
+                            },
+                          },
+                          {
+                            terms: {
+                              id: [
+                                'dit:aventri:Event:1:Create',
+                                'dit:aventri:Event:2:Create',
+                              ],
+                            },
+                          },
+                        ],
+                      },
+                    },
+                  ],
+                },
+              },
+            },
+          },
+        }
 
-      //   expect(fetchActivityFeedStub).to.be.calledWith(
-      //     middlewareParameters.reqMock,
-      //     expectedEsQuery
-      //   )
-      // })
+        expect(fetchActivityFeedStub).to.be.calledWith(
+          middlewareParameters.reqMock,
+          expectedEsQuery
+        )
+      })
     })
 
     context(
@@ -531,8 +485,7 @@ describe('Activity feed controllers', () => {
               is_global_ultimate: true,
             },
             requestQuery: {
-              activityTypeFilter: 'dataHubActivity',
-              showDnbHierarchy: true,
+              showDnbHierarchy: [true],
             },
             user: {
               id: 123,
@@ -546,89 +499,116 @@ describe('Activity feed controllers', () => {
           )
         })
 
-        // it('should call fetchActivityFeed with the right params', async () => {
-        //   const expectedEsQuery = {
-        //     from: 0,
-        //     size: 20,
-        //     sort: sortCriteria('desc'),
-        //     query: {
-        //       bool: {
-        //         filter: {
-        //           bool: {
-        //             should: [
-        //               {
-        //                 bool: {
-        //                   must: [
-        //                     {
-        //                       terms: {
-        //                         'object.type': DATA_HUB_AND_AVENTRI_ACTIVITY,
-        //                       },
-        //                     },
-        //                     {
-        //                       terms: {
-        //                         'object.attributedTo.id': [
-        //                           'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
-        //                           'dit:DataHubCompany:123',
-        //                           'dit:DataHubCompany:456',
-        //                         ],
-        //                       },
-        //                     },
-        //                   ],
-        //                 },
-        //               },
-        //               {
-        //                 bool: {
-        //                   must: [
-        //                     {
-        //                       term: {
-        //                         'object.type': 'dit:aventri:Event',
-        //                       },
-        //                     },
-        //                     {
-        //                       terms: {
-        //                         id: [
-        //                           'dit:aventri:Event:1:Create',
-        //                           'dit:aventri:Event:2:Create',
-        //                         ],
-        //                       },
-        //                     },
-        //                   ],
-        //                 },
-        //               },
-        //               {
-        //                 bool: {
-        //                   must: [
-        //                     {
-        //                       term: {
-        //                         'object.attributedTo.id':
-        //                           'dit:directoryFormsApi:SubmissionType:export-support-service',
-        //                       },
-        //                     },
-        //                     {
-        //                       terms: {
-        //                         'actor.dit:emailAddress': [
-        //                           'fred@acme.org',
-        //                           'fred@acme.org',
-        //                           'fred@acme.org',
-        //                           'byvanuwenu@yahoo.com',
-        //                         ],
-        //                       },
-        //                     },
-        //                   ],
-        //                 },
-        //               },
-        //             ],
-        //           },
-        //         },
-        //       },
-        //     },
-        //   }
+        it('should call fetchActivityFeed with the right params', async () => {
+          const expectedEsQuery = {
+            from: 0,
+            size: 20,
+            sort: sortCriteria('desc'),
+            query: {
+              bool: {
+                filter: {
+                  bool: {
+                    should: [
+                      {
+                        bool: {
+                          must: [
+                            {
+                              terms: {
+                                'object.type': DATA_HUB_AND_EXTERNAL_ACTIVITY,
+                              },
+                            },
+                            {
+                              terms: {
+                                'object.attributedTo.id': [
+                                  'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
+                                ],
+                              },
+                            },
+                          ],
+                        },
+                      },
+                      {
+                        bool: {
+                          must: [
+                            {
+                              term: {
+                                'object.type':
+                                  'dit:directoryFormsApi:Submission',
+                              },
+                            },
+                            {
+                              term: {
+                                'object.attributedTo.type':
+                                  'dit:directoryFormsApi:SubmissionAction:gov-notify-email',
+                              },
+                            },
+                            {
+                              term: {
+                                'object.url': '/contact/export-advice/comment/',
+                              },
+                            },
+                            {
+                              terms: {
+                                'actor.dit:emailAddress': [
+                                  'fred@acme.org',
+                                  'fred@acme.org',
+                                  'fred@acme.org',
+                                  'byvanuwenu@yahoo.com',
+                                ],
+                              },
+                            },
+                          ],
+                        },
+                      },
+                      {
+                        bool: {
+                          must: [
+                            { term: { 'object.type': 'dit:aventri:Event' } },
+                            {
+                              terms: {
+                                id: [
+                                  'dit:aventri:Event:1:Create',
+                                  'dit:aventri:Event:2:Create',
+                                ],
+                              },
+                            },
+                          ],
+                        },
+                      },
+                      {
+                        bool: {
+                          must: [
+                            {
+                              term: {
+                                'object.attributedTo.id':
+                                  'dit:directoryFormsApi:SubmissionType:export-support-service',
+                              },
+                            },
+                            {
+                              terms: {
+                                'actor.dit:emailAddress': [
+                                  'fred@acme.org',
+                                  'fred@acme.org',
+                                  'fred@acme.org',
+                                  'byvanuwenu@yahoo.com',
+                                ],
+                              },
+                            },
+                          ],
+                        },
+                      },
+                    ],
+                  },
+                },
+              },
+            },
+          }
 
-        //   expect(fetchActivityFeedStub).to.be.calledWith(
-        //     middlewareParameters.reqMock,
-        //     expectedEsQuery
-        //   )
-        // })
+          expect(fetchActivityFeedStub).to.be.calledWith(
+            middlewareParameters.reqMock,
+            expectedEsQuery
+          )
+        })
       }
     )
 
@@ -652,65 +632,93 @@ describe('Activity feed controllers', () => {
         )
       })
 
-      // it('should default to dataHubActivity', async () => {
-      //   const expectedEsQuery = {
-      //     from: 0,
-      //     size: 20,
-      //     sort: sortCriteria('desc'),
-      //     query: {
-      //       bool: {
-      //         filter: {
-      //           bool: {
-      //             should: [
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       terms: {
-      //                         'object.type': DATA_HUB_AND_AVENTRI_ACTIVITY,
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         'object.attributedTo.id': [
-      //                           'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //               {
-      //                 bool: {
-      //                   must: [
-      //                     {
-      //                       term: {
-      //                         'object.type': 'dit:aventri:Event',
-      //                       },
-      //                     },
-      //                     {
-      //                       terms: {
-      //                         id: [
-      //                           'dit:aventri:Event:1:Create',
-      //                           'dit:aventri:Event:2:Create',
-      //                         ],
-      //                       },
-      //                     },
-      //                   ],
-      //                 },
-      //               },
-      //             ],
-      //           },
-      //         },
-      //       },
-      //     },
-      //   }
+      it('should default to data hub and external activity', async () => {
+        const expectedEsQuery = {
+          from: 0,
+          size: 20,
+          sort: sortCriteria('desc'),
+          query: {
+            bool: {
+              filter: {
+                bool: {
+                  should: [
+                    {
+                      bool: {
+                        must: [
+                          {
+                            terms: {
+                              'object.type': DATA_HUB_AND_EXTERNAL_ACTIVITY,
+                            },
+                          },
+                          {
+                            terms: {
+                              'object.attributedTo.id': [
+                                'dit:DataHubCompany:dcdabbc9-1781-e411-8955-e4115bead28a',
+                              ],
+                            },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      bool: {
+                        must: [
+                          {
+                            term: {
+                              'object.type': 'dit:directoryFormsApi:Submission',
+                            },
+                          },
+                          {
+                            term: {
+                              'object.attributedTo.type':
+                                'dit:directoryFormsApi:SubmissionAction:gov-notify-email',
+                            },
+                          },
+                          {
+                            term: {
+                              'object.url': '/contact/export-advice/comment/',
+                            },
+                          },
+                          {
+                            terms: {
+                              'actor.dit:emailAddress': [
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'fred@acme.org',
+                                'byvanuwenu@yahoo.com',
+                              ],
+                            },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      bool: {
+                        must: [
+                          { term: { 'object.type': 'dit:aventri:Event' } },
+                          {
+                            terms: {
+                              id: [
+                                'dit:aventri:Event:1:Create',
+                                'dit:aventri:Event:2:Create',
+                              ],
+                            },
+                          },
+                        ],
+                      },
+                    },
+                  ],
+                },
+              },
+            },
+          },
+        }
 
-      //   expect(fetchActivityFeedStub).to.be.calledWith(
-      //     middlewareParameters.reqMock,
-      //     expectedEsQuery
-      //   )
-      // })
+        expect(fetchActivityFeedStub).to.be.calledWith(
+          middlewareParameters.reqMock,
+          expectedEsQuery
+        )
+      })
     })
 
     context('when the endpoint returns error', () => {
@@ -750,7 +758,6 @@ describe('Activity feed controllers', () => {
           middlewareParameters = buildMiddlewareParameters({
             company: companyMock,
             requestQuery: {
-              activityTypeFilter: 'dataHubAndExternalActivity',
               showDnbHierarchy: false,
             },
             user: {
